wine.inf.in: Add invalid dummy certificate to CA certificate store.
Pythons ssl module raises an exception if the CA certificate store is empty.
Wine currently doesn't add any certificates to this store and this patch tries
to solve the problem by adding a dummy certificate. The certificate shouldn't
have any effect though as it is invalid by design. The dates were choosen in a
way that this certificate can never be valid:

Not Before: Mar 13 02:13:31 2016 GMT
Not After : Mar 12 02:13:31 2016 GMT