configure: Avoid clobbering x18 on arm64 within wine
On aarch64/arm64 on linux, the compiler is free to use x18 for normal
code generation (while the register is reserved on iOS/darwin, and
on windows).

If targeting arm64, check for the flags that allows this register to
be left untouched (the flag is supported both by gcc and clang).

Similar issues can still pop up as soon as system library functions
that happen to touch x18 are called, unless the system libraries have
been built with the same flag.

Signed-off-by: Martin Storsjo <martin@martin.st>