winex11.drv/window: Query the X server for the actual rect of the window before unmapping it
Some applications control their own position when they are being moved. This
can become out of sync with the X server because certain WMs interfere with
this on purpose.

Thus we shouldn't just rely on our window rect when deciding whether to unmap
a window (if it's out of the screen), but query the X server in this case
as the last resort to make sure it really is outside and should be unmapped.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=15346
Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>