winex11.drv: Report correct display orientation for the fullscreen hack.
This restores the functionality of eb55e52fc48b. Fullscreen hack reports correct display
orientation with this patch. However, it doesn't support changing display orientations. The purpose
of the fullscreen hack is not to make any display changes to the host system. So if we were to
support display orientations for the fullscreen hack, we would have to do rotation transforms in
many places and it would be messy.

Signed-off-by: Zhiyi Zhang <zzhang@codeweavers.com>