winex11.drv: Let the Window Manager handle the off-screen window if it's managed.
Some applications control their own position when they are being moved. This
means WindowPosChanged is called without an event.

Window Managers have their own way to deal with offscreen windows, and we
have to respect that, if the window is managed. So let the WM handle this
situation instead of unmapping the window ourselves.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=15346
Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>